FELIX-6155 - Add caching of service component metadata
By default the caching is disabled.  The configuration property (or
framework launch property) ds.cache.metadata can be set to the string
"true" to enable the support.  Unfortunately dynamically setting this
with config admin is problematic because the cache is read at activation
time currently.  This could be enhanced to read the cache any time the
value is set to true dynamically.  For now it is required to be set from
the start with a system or framework launch property

git-svn-id: https://svn.apache.org/repos/asf/felix/trunk@1862491 13f79535-47bb-0310-9956-ffa450edef68