WORK IN PROGRESS
Initial work to add a component.manager concept that allows bundles with
components to provide a ComponentManagerFactory which SCR uses to create
ComponentManager objects for managing a component instance
creating/modifying/injecting with the use of little reflection as
possible.

Note this is only work in progress and needs to be polished and tested.