Add CachedMethodInfo for lifecycle methods.
The algorithm to find the lifecycle methods (activate, modified,
deactivate) is expensive unless using only DS 1.0 method signatures.
This adds additional caching of the lifecycle methods so that on a
cached restart the search can do a single getDeclaredMethod call to find
the exact lifecycle method.

The code for BaseMethod was updated such that the concept can be
extended to the bind methods also.  At this time the bind methods seem
to perform fine at least for the simple single parameter methods.